Null Jenkins build result now considered success.
When a Jenkins build is in-progress, its result is null. If Dashing
polls for the result while the build is running, it will be reflected
as a failure on the dashboard (particularly noticeable in projects
with long build times).

A dashboard's purpose is to provide at-a-glance information.
Incorrectness in a context where the information drives reactive
behavior may result in wasteful activities and, over time, something
akin to alert fatigue.

This commit changes the interpretation of a "null" build result in
Jenkins from failure to success. Some questionable assumptions are
made:

1. A Jenkins build result of "null" means the build is in-progress.
2. An in-progress build should still be displayed in Dashing.
3. In-progress builds should be considered successful until proven
   otherwise.